+++ src/helpers/Config.ts
@@ -13,9 +13,17 @@ export interface ModuleData {
 }
 
 /**
+ * Base application option type with the log level property.
+ */
+export interface AppOption {
+  logLevel: 'silly' | 'trace' | 'debug' | 'info' | 'warn' | 'error' | 'fatal';
+}
+
+/**
  * Base configuration.
  */
-export interface ConfigBase<T> {
+export interface ConfigBase<T extends AppOption> {
+  logLevel: 'silly' | 'trace' | 'debug' | 'info' | 'warn' | 'error' | 'fatal' | undefined;
   mqtt: {
     host: string | undefined;
     port: number | undefined;
@@ -36,7 +44,7 @@ export interface ConfigBase<T> {
 /**
  * Config helper class.
  */
-export class Config<T> {
+export class Config<T extends AppOption> {
   private logger: Logger;
 
   private mqttOption: MqttBrokerClientOption;
@@ -51,6 +59,7 @@ export class Config<T> {
       host: Config.useValueOrDefault('mqtt.host', config.mqtt.host, 'localhost'),
       port: Config.useValueOrDefault('mqtt.port', config.mqtt.port, 1883),
       system: Config.useValueOrDefault('mqtt.system', config.mqtt.system, module.name),
+      logLevel: Config.useValueOrDefault('logLevel', config.logLevel, 'info'),
     };
     this.influxOption = {
       host: Config.useValueOrDefault('influx.host', config.influx.host, 'localhost'),
@@ -60,8 +69,10 @@ export class Config<T> {
       token: Config.useValueOrDefault('influx.token', config.influx.token, ''),
       org: Config.useValueOrDefault('influx.org', config.influx.org, 'default'),
       bucket: Config.useValueOrDefault('influx.bucket', config.influx.bucket, module.name),
+      logLevel: Config.useValueOrDefault('logLevel', config.logLevel, 'info'),
     };
     this.appOption = config.app;
+    this.appOption.logLevel = Config.useValueOrDefault('logLevel', config.logLevel, 'info');
   }
 
   /**