+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0
+
+* Added: Optional configuration for the log level
+* Added: Additional debug and trace log output
+
 ## 1.0.0
 
 * Changed: Completely reworked for the smart home with a core MQTT communication